Play with the ensemble dataset a bit
As suspected, the load time sucks :) ~10x slower in my rig for a day
without a lot of predicted rain (timing is dominated by DEFLATE, so
more zeroes in the data -> faster decompression -> faster load)

Should be a bit worse in prod, which means a load time of ~3s without
any extra load. Too high for my taste but maybe I can get away with
it. We'll see...